Updated the player character and the projectile. Added enemy, enemy projectile and updated the what to have list
Updated the player character so that it spawns in some projectiles in the beginning, so we have a small pool of rockets.
The projectiles are now reusable. We dont destroy them, only dissables them.
Added a enemy and enemy projectile to the game. The enemy has its movement and shooting ready, just need to call the functions properly. They can not yet be destroyed. Wrote some more points to the what to have list, like an enemy manager that keeps track of all the enemies.